[Broken at the moment] Change to enums
