+++ src/main/java/seedu/address/logic/parser/TabSwitchCommandParser.java
@@ -3,11 +3,12 @@ package seedu.address.logic.parser;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_TAB_NAME;
 
-import seedu.address.logic.commands.TabSwitchCommand;
+import seedu.address.commons.core.Tab;
+import seedu.address.logic.commands.TabCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.TabSwitch;
 
-public class TabSwitchCommandParser implements Parser<TabSwitchCommand> {
+public class TabSwitchCommandParser implements Parser<TabCommand> {
     /**
      * Parses {@code userInput} into a command and returns it.
      *
@@ -15,10 +16,10 @@ public class TabSwitchCommandParser implements Parser<TabSwitchCommand> {
      * @throws ParseException if {@code userInput} does not conform the expected format
      */
     @Override
-    public TabSwitchCommand parse(String args) throws ParseException {
+    public TabCommand parse(String args) throws ParseException {
         requireNonNull(args);
         try {
-            return new TabSwitchCommand(TabSwitch.Tab.toEnum(args.trim()));
+            return new TabCommand(Tab.toEnum(args.trim()));
         } catch (ParseException pe) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_TAB_NAME, args),
