+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -8,7 +8,7 @@ import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.RELATIONSHIP_DESC_AMY;
-import static seedu.address.model.TabSwitch.Tab.CONTACTS;
+import static seedu.address.commons.core.Tab.CONTACTS;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.AMY;
